Ensured sushi unit tests work with buffer sizes 64 & 512

+++ src/engine/transport.cpp
@@ -245,9 +245,8 @@ void Transport::_update_internals()
 
 void Transport::_update_internal_sync(int64_t samples)
 {
-    /* Assume that if there are missed callbacks, the numbers of samples
-     * will still be a multiple of AUDIO_CHUNK_SIZE */
-    auto chunks_passed = samples / AUDIO_CHUNK_SIZE;
+    // We cannot assume chunk size is an absolute multiple of samples for all buffer sizes.
+    auto chunks_passed = static_cast<float>(samples) / AUDIO_CHUNK_SIZE;
 
     if (_playmode != _set_playmode)
     {
